requirements.txt
streamlit
vtk>=9.4.0
pyvista
stpyvista
cadquery
numpy
panel
pythonocc-core

